Add resizable layout for guest list
We have an issue with the guest list sizing. It's far too small. The main area, where the guest id input field and write tag button lives is taking up around 70% of the screen real estate. The guest list must be taller. Can we make the sections resizable inline?

```git-revs
4a92524  (Base revision)
8fc4654  Snapshot before codemcp change
8ed287e  Snapshot before codemcp change
```

codemcp-id: 4-add-resizable-layout-for-guest-list